support daylight savings time
The old strategy of using tzutil on windows turned out to be invalid,
since the timestamps emitted by that tool are specific to the time of
year that you run it apparently.

This went a long time without noticing because most calendar events are
in US timezones and daylight savings time is coordinated between them,
so there was no relative errors.